Add files via upload
This update of KCAlgoBase, ATMAlgoBase and the main bots with and without ATM set the default settings to RegChan profit target and trailing stop.  You now can select "Entries Per Direction."  It was in the algo base without being exposed.  The 5 time slots are configured based on the bots' live sim performance.  I will narrow down only the profitable timeframes as I gather more data.  The timeframes are now in EST time zone